switched allFactors() to returning long

+++ euler.h
@@ -1,5 +1,5 @@
 int multi35(); // Problem 1
 
-int allFactors(); // Problem 5
+long allFactors(); // Problem 5
 
 int sumSquareDiff(); // Problem 6
